requirements.txt
flask
requests
torch
transformers
pillow
